#33 refactor(ticketing): 티켓 발행 요청 DTO 수정 및 패키지 구조 재구성
