+++ payment/src/main/java/wisoft/nextframe/payment/infra/payment/adaptor/TossPaymentsAdaptor.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.payment.infra.payment.client;
+package wisoft.nextframe.payment.infra.payment.adaptor;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
