+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/ticketing/TicketIssueRequest.java
@@ -1,0 +1,6 @@
+package wisoft.nextframe.schedulereservationticketing.dto.ticketing;
+
+import java.util.UUID;
+
+public record TicketIssueRequest(UUID reservationId ) {
+}
