+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/exception/ticketing/CannotIssueTicketWithoutPaymentException.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.schedulereservationticketing.ticketing.exception;
+package wisoft.nextframe.schedulereservationticketing.exception.ticketing;
 
 public class CannotIssueTicketWithoutPaymentException extends RuntimeException {
 	public CannotIssueTicketWithoutPaymentException() {
