+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/repository/ticketing/TicketRepository.java
@@ -1,11 +1,11 @@
-package wisoft.nextframe.schedulereservationticketing.ticketing.repository;
+package wisoft.nextframe.schedulereservationticketing.repository.ticketing;
 
 import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import wisoft.nextframe.schedulereservationticketing.ticketing.entity.Ticket;
+import wisoft.nextframe.schedulereservationticketing.entity.ticketing.Ticket;
 
 public interface TicketRepository extends JpaRepository<Ticket, UUID> {
 	Optional<Ticket> findByReservationId(UUID reservationId);
