+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/service/ticketing/PaymentClient.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.schedulereservationticketing.ticketing.service;
+package wisoft.nextframe.schedulereservationticketing.service.ticketing;
 
 import java.util.UUID;
