--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/ticketing/controller/dto/TicketIssueRequest.java
@@ -1,6 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.ticketing.controller.dto;
-
-import java.util.UUID;
-
-public record TicketIssueRequest(UUID reservationId, UUID paymentId) {
-}
