+++ schedule-reservation-ticketing/src/main/resources/application.yaml
@@ -43,3 +43,6 @@ logging:
       max-file-size: 10MB # 파일당 최대 용량 KB, MB, GB, 10MB를 넘어가면 %i 숫자 증가되면서 롤링
       max-history: 60
       total-size-cap: 1GB
+
+server:
+  port: 8081
